+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/feature/messages/list/adapter/MessageListItemViewType.kt
@@ -39,4 +39,5 @@ public object MessageListItemViewType {
     public const val FILE_ATTACHMENTS: Int = OFFSET + 14
     public const val LINK_ATTACHMENTS: Int = OFFSET + 15
     public const val UNREAD_SEPARATOR: Int = OFFSET + 16
+    public const val START_OF_THE_CHANNEL: Int = OFFSET + 17
 }